Add request param required property update checker
