+++ checker/checker_breaking_test.go
@@ -184,30 +184,6 @@ func TestBreaking_NewRequiredHeaderParam(t *testing.T) {
 	require.Equal(t, "new-required-request-parameter", errs[0].Id)
 }
 
-// BC: changing an existing header param from optional to required is breaking
-func TestBreaking_HeaderParamRequiredEnabled(t *testing.T) {
-	s1 := l(t, 1)
-	s2 := l(t, 1)
-
-	s1.Spec.Paths[installCommandPath].Get.Parameters.GetByInAndName(openapi3.ParameterInHeader, "network-policies").Required = false
-	s2.Spec.Paths[installCommandPath].Get.Parameters.GetByInAndName(openapi3.ParameterInHeader, "network-policies").Required = true
-
-	d, osm, err := diff.GetWithOperationsSourcesMap(getConfig(), &s1, &s2)
-	require.NoError(t, err)
-	errs := checker.CheckBackwardCompatibility(checker.GetDefaultChecks(), d, osm)
-	require.NotEmpty(t, errs)
-	require.Equal(t, checker.BackwardCompatibilityErrors{
-		{
-			Id:        "request-parameter-became-required",
-			Text:      "the 'header' request parameter 'network-policies' became required",
-			Comment:   "",
-			Level:     checker.ERR,
-			Operation: "GET",
-			Path:      "/api/{domain}/{project}/install-command",
-			Source:    "../data/openapi-test1.yaml",
-		}}, errs)
-}
-
 // BC: changing an existing response header from required to optional is breaking
 func TestBreaking_ResponseHeaderParamRequiredDisabled(t *testing.T) {
 	s1 := l(t, 1)
